Actualizacion general de archivo de la aplicacion y css.

+++ application/views/back_end/comen_alum_view.php
@@ -22,7 +22,7 @@ if ($this->session->userdata('nombre')) {
     <section class="contenido">
         <script type="text/javascript" src="<?= base_url() ?>/assets/moment.js"></script>
         <h2><?= $entry->tar_descripcion ?></h2>
-         Fecha de Creación: <?= $entry->tar_fechaasignacion ?><p></p>
+        Fecha de Creación: <?= $entry->tar_fechaasignacion ?><p></p>
         Fecha de entrega: <?= $entry->tar_fechaentrega ?><p></p>
         Archivo Adjunto: <?= $entry->tar_nombrearchivo ?><p></p>
         <br />
@@ -30,23 +30,54 @@ if ($this->session->userdata('nombre')) {
 
         <?php if ($this->session->userdata('nombre')) : ?>
             Tu comentario: 
-            <?= form_open(base_url() . 'backend/profesor_control/comentario/') ?>
-            <?= form_hidden('idtarea', $this->uri->segment(4)) ?>
-            <?= form_textarea('comentario') ?>
-            <?= form_submit('submit', 'Enviar') ?>
-            <?= form_close() ?>
+            <form action="<?php echo base_url(); ?>backend/alumnos_control/comentario/" method="post" enctype="multipart/form-data" >
+                <?= form_hidden('idtarea', $this->uri->segment(4)) ?>
+                <?= form_textarea('comentario') ?>
+                <p><label for="userfile">Adjuntar Archivo:</label>
+                    <input type="file" name="userfile"  id="userfile"/>
+                    <font color='red' style='font-weight: bold; font-size: 14px; text-decoration: underline'><?php
+                    if (isset($error)) {
+                        echo $error;
+                    }
+                    ?></font>
+                    <?= form_submit('submit', 'Enviar') ?>
+            </form>
         <?php endif; ?>
-        <P ALIGN=RIGHT><a href="<?php echo base_url() ?>backend/profesor_control/vertareas">Volver a Tareas</a></p>
+        <P ALIGN=RIGHT><a href="<?php echo base_url() ?>backend/alumnos_control/successalumnostareas">Volver a Tareas</a></p>
         <?php
         if (!empty($comments)) {
             echo '<h3>Comentarios</h3>';
             foreach ($comments as $comment)
-                echo '<h4>' . $comment->autor . '</h4>' .
-                $comment->comentario . '<br/> <p>'  .
-                convertDateTimetoTimeAgo($comment->fecha) . '<hr />';
-        } else
-            echo '<h3>Sin Comentarios!</h3>';
-        ?>
+                if ($comment->autor == $this->session->userdata('nombre') and($comment->ruta_archivo)=='NULL') {
+                    echo '<h4>' . $comment->autor . '</h4>' .
+                    $comment->comentario . '<br/> <p>' .
+                    convertDateTimetoTimeAgo($comment->fecha)
+                    ?>
+                    <p><a href="<?php echo base_url() . 'backend/alumnos_control/eliminarcomentario/' . $comment->idcomentario ?>" onclick="return confirm('¿Estás seguro que desea eliminar este Comentario?')">Eliminar Comentario</a>
+                        <?php
+                    } elseif ($comment->autor != $this->session->userdata('nombre') and ($comment->ruta_archivo)=='NULL') {
+                        echo '<h4>' . $comment->autor . '</h4>' .
+                        $comment->comentario . '<br/> <p>' .
+                        convertDateTimetoTimeAgo($comment->fecha) . '<hr />';
+                    } elseif ($comment->autor != $this->session->userdata('nombre') and ($comment->ruta_archivo)!='NULL') {
+                        echo '<h4>' . $comment->autor . '</h4>' .
+                        $comment->comentario . '<br/> <p>' .
+                        convertDateTimetoTimeAgo($comment->fecha)
+                        ?>
+                    <p><a href="<?php echo base_url() . 'backend/alumnos_control/descargar/' . $comment->idcomentario ?>">Descargar Adjunto</a>
+                        <?php
+                    }elseif ($comment->autor == $this->session->userdata('nombre')and($comment->ruta_archivo)!='NULL') {
+                        echo '<h4>' . $comment->autor . '</h4>' .
+                        $comment->comentario . '<br/> <p>' .
+                        convertDateTimetoTimeAgo($comment->fecha)
+                        ?>
+                         <p aling="right"><a href="<?php echo base_url() . 'backend/alumnos_control/eliminarcomentario/' . $comment->idcomentario ?>" onclick="return confirm('¿Estás seguro que desea eliminar este Comentario?')">Eliminar Comentario</a>
+                         <p aling="left"><a href="<?php echo base_url() . 'backend/alumnos_control/descargar/' . $comment->idcomentario ?>">Descargar Adjunto</a>
+                        <?php
+                    }
+            } else
+                echo '<h3>Sin Comentarios!</h3>';
+            ?>
     </section>
 
     <?php
